VMO-2623 move this.setResourceData() to BaseMounted
