+++ stories/LocationResponseBlock.stories.ts
@@ -77,9 +77,7 @@ export const ExistingDataPreFilled = () => (CurrentClass2)
           // @ts-ignore
             const {block: {uuid: blockId}, flow: {uuid: flowId}} = await baseMounted.bind(this)(BLOCK_TYPE, locationResponseBlockStore)
 
-            this.block_setName({blockId: blockId, value: "A Name"})
-            this.block_setLabel({blockId: blockId, value: "A Label"})
-            this.block_setSemanticLabel({blockId: blockId, value: "A Semantic Label"})
+          this.setDescription(blockId)
 
             // Fake a 1st block to make sure the current block won't be selected
             // @ts-ignore
