+++ stories/NumericResponseBlock.stories.ts
@@ -64,54 +64,18 @@ export const Default = () => (CurrentClass1)
   async mounted() {
     // @ts-ignore
     const {block: {uuid: blockId}, flow: {uuid: flowId}} = await baseMounted.bind(this)(BLOCK_TYPE, numericResponseBlockStore)
-    const {
-      languages: {
-        0: {id: languageId}
-      },
-    }: IFlow = this.activeFlow
-    const resourceId = get(this.activeBlock, `config.prompt`, '')
 
     this.setDescription(blockId)
-    this.setResourceData(languageId, resourceId)
+    this.setResourceData({
+      shouldSetChoices: false,
+      configPath: 'config.prompt'
+    })
     this.setValidationMinimum({blockId, value:0})
     this.setValidationMaximum({blockId, value:99})
     this.setMaxDigits({blockId, value:2})
   },
 })
 class CurrentClass2 extends BaseMountedVueClassWithResourceAndMode {
-
-  setDescription(blockId: string) { // TODO: Find a wait to define this in BaseClass or other ParentClass without '_this.setDescription is not a function' error
-    this.block_setName({blockId: blockId, value: "A Name"})
-    this.block_setLabel({blockId: blockId, value: "A Label"})
-    this.block_setSemanticLabel({blockId: blockId, value: "A Semantic Label"})
-  }
-
-  setResourceData(languageId: string, resourceId: string) {
-    // Set values on resource editor // TODO: find better way to do this once the resource editor is fully implemented
-    const variantSms: IResourceDefinitionVariantOverModesFilter = {
-      languageId,
-      modes: [SupportedMode.SMS],
-      // @ts-ignore: TODO: remove this ts-ignore once we find a way to match `contentType` type from /@floip/flow-runner/dist/domain/IResourceResolver.d.ts:IResourceDefinitionContentTypeSpecific interface
-      contentType: [SupportedContentType.TEXT],
-    }
-    const variantUssd: IResourceDefinitionVariantOverModesFilter = {
-      languageId,
-      modes: [SupportedMode.USSD],
-      // @ts-ignore: TODO: remove this ts-ignore once we find a way to match `contentType` type from /@floip/flow-runner/dist/domain/IResourceResolver.d.ts:IResourceDefinitionContentTypeSpecific interface
-      contentType: [SupportedContentType.TEXT],
-    }
-    const variantIvr: IResourceDefinitionVariantOverModesFilter = {
-      languageId,
-      modes: [SupportedMode.IVR],
-      // @ts-ignore: TODO: remove this ts-ignore once we find a way to match `contentType` type from /@floip/flow-runner/dist/domain/IResourceResolver.d.ts:IResourceDefinitionContentTypeSpecific interface
-      contentType: [SupportedContentType.AUDIO],
-    }
-    // we're assuming this pseudo-variants exist
-    this.resource_setValue({resourceId, filter: variantSms, value: "text for SMS"})
-    this.resource_setValue({resourceId, filter: variantUssd, value: "text for USSD"})
-    this.resource_setValue({resourceId, filter: variantIvr, value: "path/to/ivr audio.mp3"})
-  }
-
   @blockVuexNamespace.Action setValidationMinimum:any
   @blockVuexNamespace.Action setValidationMaximum:any
   @blockVuexNamespace.Action setMaxDigits:any
@@ -125,54 +89,19 @@ export const ExistingDataForAllModes = () => (CurrentClass2)
   async mounted() {
     // @ts-ignore
     const {block: {uuid: blockId}, flow: {uuid: flowId}} = await baseMounted.bind(this)(BLOCK_TYPE, numericResponseBlockStore)
-    const {
-      languages: {
-        0: {id: languageId}
-      },
-    }: IFlow = this.activeFlow
-    const resourceId = get(this.activeBlock, `config.prompt`, '')
 
     this.flow_setSupportedMode({flowId, value: SupportedMode.IVR})
     this.setValidationMinimum({blockId, value:0})
     this.setValidationMaximum({blockId, value:99})
     this.setMaxDigits({blockId, value:2})
     this.setDescription(blockId)
-    this.setResourceData(languageId, resourceId)
+    this.setResourceData({
+      shouldSetChoices: false,
+      configPath: 'config.prompt'
+    })
   },
 })
 class CurrentClass3 extends BaseMountedVueClassWithResourceAndMode {
-  setDescription(blockId: string) { // TODO: Find a wait to define this in BaseClass or other ParentClass without '_this.setDescription is not a function' error
-    this.block_setName({blockId: blockId, value: "A Name"})
-    this.block_setLabel({blockId: blockId, value: "A Label"})
-    this.block_setSemanticLabel({blockId: blockId, value: "A Semantic Label"})
-  }
-
-  setResourceData(languageId: string, resourceId: string) {
-    // Set values on resource editor // TODO: find better way to do this once the resource editor is fully implemented
-    const variantSms: IResourceDefinitionVariantOverModesFilter = {
-      languageId,
-      modes: [SupportedMode.SMS],
-      // @ts-ignore: TODO: remove this ts-ignore once we find a way to match `contentType` type from /@floip/flow-runner/dist/domain/IResourceResolver.d.ts:IResourceDefinitionContentTypeSpecific interface
-      contentType: [SupportedContentType.TEXT],
-    }
-    const variantUssd: IResourceDefinitionVariantOverModesFilter = {
-      languageId,
-      modes: [SupportedMode.USSD],
-      // @ts-ignore: TODO: remove this ts-ignore once we find a way to match `contentType` type from /@floip/flow-runner/dist/domain/IResourceResolver.d.ts:IResourceDefinitionContentTypeSpecific interface
-      contentType: [SupportedContentType.TEXT],
-    }
-    const variantIvr: IResourceDefinitionVariantOverModesFilter = {
-      languageId,
-      modes: [SupportedMode.IVR],
-      // @ts-ignore: TODO: remove this ts-ignore once we find a way to match `contentType` type from /@floip/flow-runner/dist/domain/IResourceResolver.d.ts:IResourceDefinitionContentTypeSpecific interface
-      contentType: [SupportedContentType.AUDIO],
-    }
-    // we're assuming this pseudo-variants exist
-    this.resource_setValue({resourceId, filter: variantSms, value: "text for SMS"})
-    this.resource_setValue({resourceId, filter: variantUssd, value: "text for USSD"})
-    this.resource_setValue({resourceId, filter: variantIvr, value: "path/to/ivr audio.mp3"})
-  }
-
   @blockVuexNamespace.Action setValidationMinimum:any
   @blockVuexNamespace.Action setValidationMaximum:any
   @blockVuexNamespace.Action setMaxDigits:any
@@ -186,53 +115,18 @@ export const ExistingDataForIvrOnly = () => (CurrentClass3)
   async mounted() {
     // @ts-ignore
     const {block: {uuid: blockId}, flow: {uuid: flowId}} = await baseMounted.bind(this)(BLOCK_TYPE, numericResponseBlockStore)
-    const {
-      languages: {
-        0: {id: languageId}
-      },
-    }: IFlow = this.activeFlow
-    const resourceId = get(this.activeBlock, `config.prompt`, '')
 
     this.flow_setSupportedMode({flowId, value: [SupportedMode.SMS, SupportedMode.USSD]})
     this.setValidationMinimum({blockId, value:0})
     this.setValidationMaximum({blockId, value:99})
     this.setDescription(blockId)
-    this.setResourceData(languageId, resourceId)
+    this.setResourceData({
+      shouldSetChoices: false,
+      configPath: 'config.prompt'
+    })
   },
 })
 class CurrentClass4 extends BaseMountedVueClassWithResourceAndMode {
-  setDescription(blockId: string) { // TODO: Find a wait to define this in BaseClass or other ParentClass without '_this.setDescription is not a function' error
-    this.block_setName({blockId: blockId, value: "A Name"})
-    this.block_setLabel({blockId: blockId, value: "A Label"})
-    this.block_setSemanticLabel({blockId: blockId, value: "A Semantic Label"})
-  }
-
-  setResourceData(languageId: string, resourceId: string) {
-    // Set values on resource editor // TODO: find better way to do this once the resource editor is fully implemented
-    const variantSms: IResourceDefinitionVariantOverModesFilter = {
-      languageId,
-      modes: [SupportedMode.SMS],
-      // @ts-ignore: TODO: remove this ts-ignore once we find a way to match `contentType` type from /@floip/flow-runner/dist/domain/IResourceResolver.d.ts:IResourceDefinitionContentTypeSpecific interface
-      contentType: [SupportedContentType.TEXT],
-    }
-    const variantUssd: IResourceDefinitionVariantOverModesFilter = {
-      languageId,
-      modes: [SupportedMode.USSD],
-      // @ts-ignore: TODO: remove this ts-ignore once we find a way to match `contentType` type from /@floip/flow-runner/dist/domain/IResourceResolver.d.ts:IResourceDefinitionContentTypeSpecific interface
-      contentType: [SupportedContentType.TEXT],
-    }
-    const variantIvr: IResourceDefinitionVariantOverModesFilter = {
-      languageId,
-      modes: [SupportedMode.IVR],
-      // @ts-ignore: TODO: remove this ts-ignore once we find a way to match `contentType` type from /@floip/flow-runner/dist/domain/IResourceResolver.d.ts:IResourceDefinitionContentTypeSpecific interface
-      contentType: [SupportedContentType.AUDIO],
-    }
-    // we're assuming this pseudo-variants exist
-    this.resource_setValue({resourceId, filter: variantSms, value: "text for SMS"})
-    this.resource_setValue({resourceId, filter: variantUssd, value: "text for USSD"})
-    this.resource_setValue({resourceId, filter: variantIvr, value: "path/to/ivr audio.mp3"})
-  }
-
   @blockVuexNamespace.Action setValidationMinimum:any
   @blockVuexNamespace.Action setValidationMaximum:any
 }
@@ -247,9 +141,7 @@ export const ExistingDataForTextOnly = () => (CurrentClass4)
         // @ts-ignore
         const {block: {uuid: blockId}, flow: {uuid: flowId}} = await baseMounted.bind(this)(BLOCK_TYPE, numericResponseBlockStore)
 
-        this.block_setName({blockId: blockId, value: "A Name"})
-        this.block_setLabel({blockId: blockId, value: "A Label"})
-        this.block_setSemanticLabel({blockId: blockId, value: "A Semantic Label"})
+        this.setDescription(blockId)
 
         // Fake a 1st block to make sure the current block won't be selected
         // @ts-ignore
